[OG-138] Refactor
