--- polymorphia-frontend/components/course-groups/modals/delete-animal/index.css
@@ -1,14 +1,0 @@
-@import "tailwindcss";
-@import "../../../../styles/globals.css";
-
-.delete-animal {
-  @apply w-lg flex flex-col gap-y-4;
-
-  h1 {
-    @apply text-3xl;
-  }
-}
-
-.delete-animal-buttons {
-  @apply w-full flex gap-x-4;
-}
