--- polymorphia-frontend/components/course-groups/modals/delete-course-group/index.css
@@ -1,14 +1,0 @@
-@import "tailwindcss";
-@import "../../../../styles/globals.css";
-
-.delete-course-group {
-  @apply w-lg flex flex-col gap-y-4;
-
-  h1 {
-    @apply text-3xl;
-  }
-}
-
-.delete-course-group-buttons {
-  @apply w-full flex gap-x-4;
-}
