+++ polymorphia-frontend/components/course-groups/modals/edit-course-group/index.css
@@ -1,27 +1,7 @@
 @import "tailwindcss";
 @import "../../../../styles/globals.css";
 
-.edit-course-group {
-  @apply flex flex-col gap-y-4 w-sm;
-
-  h1 {
-    @apply text-3xl;
-  }
-}
-
-.edit-course-group-columns {
-  @apply flex justify-between items-center;
-
-  h1 {
-    @apply max-w-44 truncate;
-  }
-}
-
 .edit-course-group-input {
   @apply w-40 border-b-2 border-primary-dark dark:border-secondary-gray bg-transparent text-center text-2xl pb-1;
   @apply focus:outline-none;
-}
-
-.edit-course-group-buttons {
-  @apply w-full flex gap-x-4;
 }
